+++ socket.h
@@ -1,6 +1,7 @@
 #ifndef SOCKET_H
 #define SOCKET_H
 
+#include <cstdint>
 #include <cstdio>
 #include <errno.h>
 #include <unistd.h>
@@ -28,7 +29,7 @@ class Socket{
 		int Bind(long);
 		Socket * Accept();
 		int Sendto(const char *, int, const char *, int);
-		int Sendto(const char *, int, struct in_addr, int);
+		int Sendto(const char *, int, struct in_addr, uint16_t);
 		int Sendto(const char *, int, struct sockaddr_in *);
 		int Recvfrom(char *, int, struct sockaddr_in* = nullptr);
 		int Shutdown(int);
